[refactor] move email duplication check API to EmailController
